+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Mailbox attributes [#134]
+
 ### Changed
 
 - Get signature from file [#135]
@@ -243,6 +247,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#131]: https://github.com/soywod/himalaya/issues/131
 [#132]: https://github.com/soywod/himalaya/issues/132
 [#133]: https://github.com/soywod/himalaya/issues/133
+[#134]: https://github.com/soywod/himalaya/issues/134
 [#135]: https://github.com/soywod/himalaya/issues/135
 [#136]: https://github.com/soywod/himalaya/issues/136
 [#138]: https://github.com/soywod/himalaya/issues/138